refactor(self_improve/cycle): merge inline tests into tests_cycle.rs (#1266)
Inline `#[cfg(test)] mod tests` block (lines 270-402) in cycle.rs is moved
into the existing sibling tests_cycle.rs file as a nested submodule
`cycle_inline` to scope its helper functions (make_score, dims) and avoid
collision with the existing tests_cycle.rs helpers of the same name.

- src/self_improve/cycle.rs: 402 → 269 LOC (under 400-line rule)
- src/self_improve/tests_cycle.rs: 670 → 810 LOC

Pure refactor: zero logic changes, all 42 self_improve::tests_cycle tests
pass unmodified.

Closes 1/70 of #1266.
